+++ src/clojure/lazytest/config.clj
@@ -1,28 +1,23 @@
 (ns lazytest.config
   (:require
    [clojure.java.io :as io]
-   [clojure.string :as str]))
+   [clojure.string :as str]
+   [lazytest.hooks :refer [run-hooks]]))
 
 (set! *warn-on-reflection* true)
 
 (def version (delay (str/trim (slurp (io/resource "LAZYTEST_VERSION")))))
 (defn lazytest-version [] (str "Lazytest v" @version))
 
-(defn combine-reporters
-  ([reporter] (fn [config m] (reporter config m) (flush) nil))
-  ([reporter & reporters]
-   (fn [config m]
-     (run! (fn [reporter] (reporter config m) (flush) nil)
-           (cons reporter reporters)))))
-
 ;; inspired by kaocha.config/resolve-reporter
 (defn resolve-reporter
   [reporter]
   (cond
-    (sequential? reporter) (->> (flatten reporter)
+    (sequential? reporter) (->> reporter
                                 (map resolve-reporter)
+                                (flatten)
                                 (distinct)
-                                (apply combine-reporters))
+                                (vec))
     (qualified-symbol? reporter)
     (if-let [r (requiring-resolve reporter)]
       (resolve-reporter (var-get r))
@@ -31,14 +26,41 @@
     (symbol? reporter) (resolve-reporter (symbol "lazytest.reporters" (name reporter)))
     :else reporter))
 
+(defn resolve-hooks
+  [hook]
+  (cond
+    (sequential? hook) (->> hook
+                            (map resolve-hooks)
+                            (flatten)
+                            (distinct)
+                            (vec))
+    (qualified-symbol? hook)
+    (if-let [h (requiring-resolve hook)]
+      (resolve-hooks (var-get h))
+      (throw (ex-info (str "Cannot find hook: " hook)
+                      {:hook hook})))
+    (symbol? hook) (resolve-hooks (symbol "lazytest.hooks" (name hook)))
+    :else hook))
+
+(defn to-seq [obj]
+  (cond (sequential? obj) (vec (distinct obj))
+        (nil? obj) []
+        :else [obj]))
+
 (defn ->config [config]
-  (if (:lazytest.runner/depth config)
+  (if (::generated config)
     config
-    (let [output (or (:output config) 'lazytest.reporters/nested)
-          output (if (sequential? output) (distinct output) [output])
-          reporter (resolve-reporter output)]
-      (-> config
-          (assoc :lazytest.runner/depth 1)
-          (assoc :lazytest.runner/suite-history [])
-          (assoc :output output)
-          (assoc :reporter reporter)))))
+    (let [output (to-seq (or (:output config) 'lazytest.reporters/nested))
+          reporters (resolve-reporter output)
+          hooks (to-seq (:hooks config))
+          hooks (resolve-hooks hooks)
+          config (-> config
+                     (assoc ::generated true)
+                     (assoc :lazytest.runner/depth 1)
+                     (assoc :lazytest.runner/suite-history [])
+                     (assoc :output output)
+                     (assoc :reporters reporters)
+                     (assoc :hooks hooks))]
+      (if-some [config (run-hooks config config :config)]
+        config
+        config))))